rename the model parameter dictionary file errors (it now just contains those)
