+++ landlab/core/errors.py
